NotificationController refactor
