+++ site/app/templates/NotificationSettings.twig
@@ -4,7 +4,7 @@
     }
 </style>
 <div class="content">
-    <form method="post" action="{{ core.buildUrl({'component': 'notification_settings', 'page': 'alter_notifcation_settings'}) }}" id="form_notification_settings">
+    <form method="post" action="{{ core.buildUrl({'component': 'notification', 'page': 'alter_notifcation_settings'}) }}" id="form_notification_settings">
         <div id="config">
             <div><h1>Notification/Email Settings</h1></div>
             <div class="row">
